use common folder for cohere
